requirements.txt
Flask>=2.0
pandas>=1.3
python-pptx>=0.6.21
flask-cors>=3.0
